+++ builder/yandex/step_create_instance.go
@@ -205,14 +205,17 @@ func (s *StepCreateInstance) Run(ctx context.Context, state multistep.StateBag) 
 		return stepHaltWithError(state, fmt.Errorf("Error preparing instance metadata: %s", err))
 	}
 
-	// TODO make part metadata prepare process
 	if config.UseIPv6 {
-		// this ugly hack will replace user provided 'user-data'
-		userData := `#cloud-config
-runcmd:
-- [ sh, -c, '/sbin/dhclient -6 -D LL -nw -pf /run/dhclient_ipv6.eth0.pid -lf /var/lib/dhcp/dhclient_ipv6.eth0.leases eth0' ]
-`
-		instanceMetadata["user-data"] = userData
+		ui.Say("Prepare user-data...")
+
+		oldUserData, ok := instanceMetadata["user-data"]
+		if !ok {
+			oldUserData = ""
+		}
+		instanceMetadata["user-data"], err = MergeCloudUserMetaData(oldUserData, cloudInitIPv6Config)
+		if err != nil {
+			return stepHaltWithError(state, fmt.Errorf("Error merge user data configs: %s", err))
+		}
 	}
 
 	req := &compute.CreateInstanceRequest{